// src/integrations/supabase/types.ts
export type Database = {
  // Allows to automatically instantiate createClient with right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: "14.5"
  }
  public: {
    Tables: {
// ...
      projects: {
        Row: {
          after_images: string[]
          before_images: string[]
          category: string
          created_at: string
          description: string
          duration: string
          featured: boolean
          gallery: string[]
          hero_image: string
          id: string
          location: string
          services: string[]
          title: string
        }
        Insert: {
          after_images?: string[]
          before_images?: string[]
          category?: string
          created_at?: string
          description?: string
          duration?: string
          featured?: boolean
          gallery?: string[]
          hero_image?: string
          id?: string
          location?: string
          services?: string[]
          title: string
        }
        Update: {
          after_images?: string[]
          before_images?: string[]
          category?: string
          created_at?: string
          description?: string
          duration?: string
          featured?: boolean
          gallery?: string[]
          hero_image?: string
          id?: string
          location?: string
          services?: string[]
          title?: string
        }
        Relationships: []
      }
// ...
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      has_role: {
        Args: {
          _role: Database["public"]["Enums"]["app_role"]
          _user_id: string
        }
        Returns: boolean
      }
    }
    Enums: {
      app_role: "admin" | "user"
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}